+++ config.js
@@ -51,7 +51,8 @@ const ConfigOptions = () => {
          * Imported assets related variables
          */
         importedAssets: {
-            importedAssetsDir: 'imported-assets'
+            importedAssetsDir: 'imported-assets',
+            importedAssetsSrcGlob: 'node_modules/@justeat/**/*'
         },
 
         /**