+++ CHANGELOG.md
@@ -5,7 +5,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v4.3.0
 ------------------------------
-*August 09, 2017*
+*August 10, 2017*
 
 ### Added
 - Added `config.importedAssets` object.
@@ -15,10 +15,12 @@ v4.3.0
 - Added `watch:assets` task.
 - Unit tests added for new config and pathBuilder properties.
 - The Readme *Imported Assets config* section was added.
-
+- The Readme *Imported Assets pathBuilder* section was added.
+- The Readme was updated to document the new tasks added.
 
 ### Changed
 - Updated clean task comments.
+- The `css:bundle` task loads the `importedAssetsDistDir` into postcss, so that Fozzie modules can access assets copied there.
 
 v4.2.0
 ------------------------------